slug alanı, editör, yazar ile gönderi ilişkilenme

+++ gonderi/views.py
@@ -1,24 +1,45 @@
-from django.shortcuts import render, HttpResponse, get_object_or_404, HttpResponseRedirect,redirect
+from django.shortcuts import render, HttpResponse, get_object_or_404, HttpResponseRedirect,redirect,Http404
 from .models import Gonderi
-from .form import GonderiForm
+from .form import GonderiForm,YorumForm
 from django.contrib import messages
+from django.utils.text import slugify
 
 def gonderi_index(request):
     gonderis=Gonderi.objects.all()
     #httpresponsu render metoduna çevircez.
     return render(request, 'gonderi/index.html',{ 'gonderis': gonderis })
 
-def gonderi_detail(request,id):
-    #gonderi=Gonderi.objects.get(id=1)
+def gonderi_detail(request,slug):
+    #gonderi=Gonderi.objects.get(slug=slug)
     #return  HttpResponse('BURASI post detail sayfası')
-    gonderi=get_object_or_404(Gonderi, id=id)
-    context={
-        'gonderi':gonderi,
-    }
+    gonderi=get_object_or_404(Gonderi, slug=slug)
+
+    if request.method == "POST":
+        # GELEN BİLGİLERİ KAYDETSİN
+        form = YorumForm(request.POST or None)  # request gonderiform değişkenini POST nesnesi haline getirir.
+        if form.is_valid():
+            yorum = form.save(
+                commit=False)  # yönlendirme sayfası için geri dönen değeri gönderi adındaki değişkene atadım
+            # save içerisindeki commit false alanı; nesneyi kaydetmeden çnce zorunlu alanıbelirlemeye izin versin
+            yorum.gonderi = gonderi
+            "    "  # istek yapan kullanıcıyı yazar yaptım
+            yorum.save()  # nesne kaydı okeyle
+            return HttpResponseRedirect(gonderi.get_absolute_url())
+    else:
+        # formu kullanıcıya göster (doldurulacak şekilde)
+
+        form = YorumForm()
+        context = {
+            'gonderi': gonderi,
+            'form': form,
+        }
     return render(request,'gonderi/detail.html',context)
 
 def gonderi_create(request):
 
+    if not request.user.is_authenticated():
+        #HTTP404 İMPORT
+        return  Http404()
     form=GonderiForm()
     # form nesnesi üretildi
     context = {
@@ -30,7 +51,10 @@ def gonderi_create(request):
         form = GonderiForm(request.POST or None,
                            request.FILES or None)  # request gonderiform değişkenini POST nesnesi haline getirir.
         if form.is_valid():
-            gonderi = form.save()  # yönlendirme sayfası için geri dönen değeri gönderi adındaki değişkene atadım
+            gonderi = form.save(commit=False)  # yönlendirme sayfası için geri dönen değeri gönderi adındaki değişkene atadım
+            #save içerisindeki commit false alanı; nesneyi kaydetmeden çnce zorunlu alanıbelirlemeye izin versin
+            gonderi.kullanici=request.user #istek yapan kullanıcıyı yazar yaptım
+            gonderi.save() #nesne kaydı okeyle
             messages.success(request, 'Başarılı şekilde gonderi oluşturuldu')
             return HttpResponseRedirect(gonderi.get_absolute_url())
     else:
@@ -40,9 +64,11 @@ def gonderi_create(request):
 
     return render(request, "gonderi/form.html", context)
 
-def gonderi_update(request,id):
-
-    gonderi=get_object_or_404(Gonderi, id=id) #gonderiyi getirir
+def gonderi_update(request,slug):
+    if not request.user.is_authenticated():
+        #HTTP404 İMPORT
+        return  Http404()
+    gonderi=get_object_or_404(Gonderi, slug=slug) #gonderiyi getirir
     form=GonderiForm(request.POST or None, request.FILES or None, instance=gonderi)
     # yukarda form getirme ve bu formugetirdiğinde gönderi bilgileri ile doldurmak için sondaki parametreye gonderi eklenir
 
@@ -57,9 +83,12 @@ def gonderi_update(request,id):
     }
     return render(request, 'gonderi/form.html', context)
 
-def gonderi_delete(request,id):
+def gonderi_delete(request,slug):
+    if not request.user.is_authenticated():
+        #HTTP404 İMPORT
+        return  Http404()
     #silmek istenen gönderi nesnesi gelsin
-    gonderi=get_object_or_404(Gonderi, id=id)
+    gonderi=get_object_or_404(Gonderi, slug=slug)
 
     #getirilen nesne sil
     gonderi.delete()
